Fix source usesrc hdr_ip() parsing for IPv6
The current parsing code for source <addr> usesrc hdr_ip() hardcodes the
assumption that the header will contain an IPv4 address. If it contains
an IPv6 address, a wrong IPv4 address will be generated from it.

This fix checks if the header contains an IPv4 address, by checking if
it contains at least one dot, but no colons. Otherwise the address is
parsed as an IPv6 address.